+++ src/frontend/apps/e2e/__tests__/app-impress/doc-grid-move.spec.ts
@@ -203,7 +203,7 @@ test.describe('Doc grid move', () => {
     await expect(docsGrid.getByText(titleDoc2)).toBeVisible();
 
     const row = await getGridRow(page, titleDoc1);
-    await row.getByText(`more_horiz`).click();
+    await row.getByRole('button', { name: /Open the menu of actions/ }).click();
 
     await page.getByRole('menuitem', { name: 'Move into a doc' }).click();
 
@@ -291,7 +291,7 @@ test.describe('Doc grid move', () => {
     await expect(docsGrid.getByText(titleDoc2)).toBeVisible();
 
     const row = await getGridRow(page, titleDoc1);
-    await row.getByText(`more_horiz`).click();
+    await row.getByRole('button', { name: /Open the menu of actions/ }).click();
 
     await page.getByRole('menuitem', { name: 'Move into a doc' }).click();
 
@@ -360,7 +360,7 @@ test.describe('Doc grid move', () => {
 
     // The first user should now be able to move the doc
     await page.reload();
-    await row.getByText(`more_horiz`).click();
+    await row.getByRole('button', { name: /Open the menu of actions/ }).click();
 
     await page.getByRole('menuitem', { name: 'Move into a doc' }).click();