+++ src/frontend/apps/e2e/__tests__/app-impress/doc-header.spec.ts
@@ -603,7 +603,7 @@ test.describe('Doc Header', () => {
 
     await expect(row.getByText(duplicateTitle)).toBeVisible();
 
-    await row.getByText(`more_horiz`).click();
+    await row.getByRole('button', { name: /Open the menu of actions/ }).click();
     await page.getByRole('menuitem', { name: 'Duplicate' }).click();
     const duplicateDuplicateTitle = 'Copy of ' + duplicateTitle;
     await page.getByText(duplicateDuplicateTitle).click();
@@ -633,7 +633,7 @@ test.describe('Doc Header', () => {
         hasText: childTitle,
       });
     await child.hover();
-    await child.getByText(`more_horiz`).click();
+    await child.getByRole('button', { name: /More options/ }).click();
 
     const currentUrl = page.url();